+++ backend/src/events/ws-exception.filter.ts
